+++ packages/theme/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Patch Changes
 
+- chore: support rewrite version text
 - chore: sass warning
 - Updated dependencies
   - vitepress-plugin-pagefind@0.4.8