+++ packages/theme/CHANGELOG.md
@@ -7,7 +7,7 @@
 - chore: support rewrite version text
 - chore: sass warning
 - Updated dependencies
-  - vitepress-plugin-pagefind@0.4.8
+  - vitepress-plugin-pagefind@0.4.9
   - vitepress@1.3.1
 
 ## 0.4.7